feat: proper ui elements are activated

+++ communities/views.py
@@ -44,9 +44,11 @@ def registration_boundaries(request):
     community = get_community(request.session.get('new_community_id'))
     community.source_of_boundaries = post_data['source']
     community.name_of_boundaries = post_data['name']
+    remove_previous_boundaries = post_data.get('remove_previous', True)
 
-    # remove all boundaries in this community
-    community.boundaries.all().delete()
+    # remove all previous boundaries in this community
+    if remove_previous_boundaries:
+        community.boundaries.all().delete()
 
     # add new boundaries in this community
     for coordinates in post_data['boundaries']:
